+++ package.json
@@ -20,11 +20,8 @@
     "react-scripts": "5.0.1",
     "react-table": "^7.8.0",
     "sass": "^1.69.7",
-
     "styled-components": "^6.1.6",
-
     "swiper": "^11.0.5",
-
     "web-vitals": "^2.1.4"
   },
   "scripts": {
