+++ src/pages/address/AddressInfo.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { AddressWrap } from "../../styles/address/addressinfostyle";
+
+export const AddressInfo = () => {
+  return (
+    <AddressWrap>
+      <h2> (41937) 대구광역시 중구 중앙대로 394 제일빌딩 5F 나나빛 </h2>
+
+      <button>주소 수정/삭제</button>
+    </AddressWrap>
+  );
+};
+
+export default AddressInfo;
