+++ src/pages/cartpage/CartPage.js
@@ -1,7 +1,8 @@
 import React from "react";
-import UserInfo from "../../components/UserInfo";
+import UserInfo from "../../components/userinfo/UserInfo";
 import styled from "@emotion/styled";
 import { css } from "@emotion/react";
+import BasicLayout from "../../layouts/BasicLayout";
 
 const CartTxt = styled.div`
   display: flex;
@@ -35,36 +36,6 @@ const CartHead = styled.div`
   }
 `;
 
-const CartProduct = styled.div`
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  width: 1157px;
-  height: 96px;
-  margin: 0 auto;
-  line-height: 18px;
-  img {
-    width: 92px;
-    height: 92px;
-  }
-  h2 {
-    font-weight: 400;
-    font-size: 12px;
-  }
-  button {
-    width: 76px;
-    height: 29px;
-    background: transparent;
-    border: solid 1px #d9d9d9;
-    gap: 50px;
-  }
-  div:nth-child(1) {
-    flex-basis: 300px;
-  }
-  .product-button {
-  }
-`;
-
 const CartPage = () => {
   return (
     <div>
@@ -83,26 +54,6 @@ const CartPage = () => {
         <div>합계</div>
         <div>선택</div>
       </CartHead>
-      <CartProduct>
-        <input type="checkbox"></input>
-        <img src={process.env.PUBLIC_URL + "/assets/images/logo.svg"} />
-        <div>
-          <h2>◆정기배송◆ 30만원 상당 공기청정기 무료 증정</h2>
-          <h2>[옵션: 공기청정기 애니멀즈+필터] </h2>
-          <button>옵션변경</button>
-        </div>
-        <div>45,900</div>
-        <div>123</div>
-        <div>기본배송</div>
-        <div>무료</div>
-        <div>45,900</div>
-        <div>1</div>
-        <div>
-          <button>관심상품 등록</button>
-          <br />
-          <button>상품 삭제</button>
-        </div>
-      </CartProduct>
     </div>
   );
 };
