+++ src/styles/address/addressinfostyle.js
@@ -1,0 +1,26 @@
+import styled from "styled-components";
+
+export const AddressWrap = styled.div`
+  width: 1155px;
+  height: 112px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  border: solid 1px #d9d9d9;
+  margin: 0 auto;
+  h2 {
+    margin-left: 50px;
+    font-size: 20px;
+    font-weight: 500;
+    color: #575757;
+  }
+  button {
+    width: 211px;
+    height: 70px;
+    border: none;
+    background: #e9b25f;
+    color: #fff;
+    font-size: 20px;
+    margin-right: 22px;
+  }
+`;
